Enroll via multichallenge for smartphone containers

+++ privacyidea/lib/containerclass.py
@@ -38,7 +38,7 @@ from privacyidea.lib.log import log_with
 from privacyidea.lib.machine import is_offline_token
 from privacyidea.lib.token import (create_tokenclass_object, get_tokens, get_serial_by_otp_list,
                                    get_tokens_from_serial_or_user)
-from privacyidea.lib.tokenclass import TokenClass
+from privacyidea.lib.tokenclass import TokenClass, CHALLENGE_SESSION
 from privacyidea.lib.user import User
 from privacyidea.lib.utils import is_true
 from privacyidea.models import (TokenContainerOwner, Realm, Token, db, TokenContainerStates,
@@ -76,6 +76,13 @@ class TokenContainerClass:
         else:
             class_options = cls.options
         return class_options
+
+    @classmethod
+    def is_multi_challenge_enrollable(cls) -> bool:
+        """
+        Returns True if the container type can be enrolled during the authentication process "via multi challenge"
+        """
+        return False
 
     def set_default_option(self, key) -> str:
         """
@@ -246,6 +253,17 @@ class TokenContainerClass:
         owners = self.get_users()
         realms = [owner.realm for owner in owners]
         return realms
+
+    @property
+    def registration_state(self) -> RegistrationState:
+        """
+        Returns the registration state of the container.
+        The registration state is stored in the container info with key 'registration_state'.
+        If the key does not exist, it returns the registration state NOT_REGISTERED with the value None.
+        """
+        container_info = self.get_container_info_dict()
+        state = container_info.get(RegistrationState.get_key())
+        return RegistrationState(state)
 
     def remove_token(self, serial: str) -> bool:
         """
@@ -704,8 +722,14 @@ class TokenContainerClass:
                               f"device_model={device_model}, key={key}, container={container} ")
                     continue
 
-                # Valid challenge: delete it
-                challenge.delete()
+                if challenge.session == CHALLENGE_SESSION.ENROLLMENT:
+                    # challenge was created during the enrollment. It is still required, hence we only set the state to
+                    # answered, but not delete it.
+                    challenge.set_otp_status(True)
+                    challenge.save()
+                else:
+                    # Valid challenge: delete it
+                    challenge.delete()
                 break
             else:
                 # Delete expired challenge
@@ -900,8 +924,7 @@ class TokenContainerClass:
         # map client and server tokens
         client_serials = [token["serial"] for token in client_tokens if "serial" in token.keys()]
 
-        container_info = self.get_container_info_dict()
-        registration_state = RegistrationState(container_info.get(RegistrationState.get_key()))
+        registration_state = self.registration_state
         if registration_state == RegistrationState.ROLLOVER_COMPLETED:
             # rollover all tokens: generate new enroll info for all tokens
             missing_serials = server_token_serials
@@ -925,6 +948,7 @@ class TokenContainerClass:
                      "They can not be added during synchronization.")
 
         # Initial synchronization after registration or rollover
+        container_info = self.get_container_info_dict()
         if initial_transfer_allowed and not is_true(container_info.get(INITIALLY_SYNCHRONIZED)):
             self.update_container_info([TokenContainerInfoData(key=INITIALLY_SYNCHRONIZED, value="True",
                                                                info_type=PI_INTERNAL)])
